R6 class to assist in reformatting the data to be used in dataSDM.

Internal functions used to temporarily store data and other information before adding to dataSDM.

Methods

Public methods

dataOrganize$makeData()

dataOrganize$makeSpecies()

dataOrganize$makeMultinom()

dataOrganize$makeFormulas()

dataOrganize$makeComponents()

dataOrganize$makeLhoods()

dataOrganize$clone()

Method makeData()

Usage

dataOrganize$makeData(
  datapoints,
  datanames,
  coords,
  proj,
  marktrialname,
  paresp,
  countsresp,
  trialname,
  speciesname,
  marks,
  pointcovnames,
  markfamily,
  temporalvar
)

Method makeSpecies()

Usage

dataOrganize$makeSpecies(speciesname)

Method makeMultinom()

Usage

dataOrganize$makeMultinom(multinomVars, return, oldVars)

Method makeFormulas()

Usage

dataOrganize$makeFormulas(
  spatcovs,
  speciesname,
  paresp,
  countresp,
  marks,
  marksspatial,
  spatial,
  intercept,
  temporalname,
  markintercept,
  pointcovs,
  speciesspatial
)

Method makeComponents()

Usage

dataOrganize$makeComponents(
  spatial,
  intercepts,
  datanames,
  marks,
  speciesname,
  multinomnames,
  pointcovariates,
  covariatenames,
  covariateclass,
  marksspatial,
  marksintercept,
  temporalname,
  speciesspatial,
  numtime,
  temporalmodel
)

Method makeLhoods()

Usage

dataOrganize$makeLhoods(
  mesh,
  ips,
  paresp,
  ntrialsvar,
  markstrialsvar,
  speciesname
)

Method clone()

The objects of this class are cloneable with this method.

Usage

dataOrganize$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.