R6 class for creating a dataSDM object to be used in runModel

A data object containing the data and the relevant information about the integrated model.

Methods

Public methods

dataSDM$print()

dataSDM$plot()

dataSDM$addData()

dataSDM$addBias()

dataSDM$updateFormula()

dataSDM$changeComponents()

dataSDM$priorsFixed()

dataSDM$specifySpatial()

dataSDM$changeLink()

dataSDM$spatialBlock()

dataSDM$new()

dataSDM$samplingBias()

Method print()

Prints the datasets, their data type and the number of observations, as well as the marks and their respective families.

Usage

dataSDM$print(...)

Arguments

...

Not used.

Method plot()

Makes a plot of the region as well as the points.

Usage

dataSDM$plot(datasetNames, Species = FALSE, Boundary = TRUE, ...)

Arguments

datasetNames

Name of the datasets to plot.

Species

Should species be plotted as well? Defaults to FALSE.

Boundary

Should a boundary (created using the inlaMesh object) be used in the plot. Defaults to TRUE.

...

Not used.

Returns

A ggplot object.

Examples

\dontrun{

#Make data object
dataObj <- intModel(...)

#Plot a specific datasets
dataObj$plot(Datasets = c('dataset1', 'dataset2'))

#Plot all datasets
dataObj$plot()

#If speciesName specified in intModel, plot species
dataObj(Species = TRUE)

}

Method addData()

Function used to add additional, and possibly heterogeneous data to the dataSDM object.

Usage

dataSDM$addData(
  ...,
  responseCounts,
  responsePA,
  trialsPA,
  markNames,
  markFamily,
  pointCovariates,
  trialsMarks,
  speciesName,
  temporalName,
  Coordinates
)

Arguments

...

The datasets added into the model: should be a data.frame or a SpatialPoints* object.

responseCounts

The name of the response variable for the counts data.

responsePA

The name of the response variable for the presence absence data.

trialsPA

The name of the trials variable for the presence absence data.

markNames

The names of the marks found in the data.

markFamily

The associated distributions of the marks.

pointCovariates

The additional, non-spatial covariates describing the data.

trialsMarks

The name of the trials variable for the binomial marks.

speciesName

The name of the species variable included in the data.

Coordinates

A vector of length 2 describing the names of the coordinates of the data.

Examples

\dontrun{

#Make data object with specified components
dataObj <- intModel(...)

#Add new data with non-standardized names
dataObj$addData(dataset, responseCounts = responseCounts,
                responsePA = responsePA, trialsPA = trialsPA,
                markNames = markNames, markFamily = markFamily,
                pointCovariates = pointCovariates,
                speciesName = speciesName, Coordinates = Coordinates)

}

Method addBias()

Function used to add additional bias fields to different data processes.

Usage

dataSDM$addBias(datasetNames = NULL, allPO = FALSE, biasField = NULL)

Arguments

datasetNames

A vector of datasets to add the bias fields to.

allPO

Should the bias fields be added to all the presence only datasets. Defaults to NULL.

biasField

An inla.spde model descrbing the random bias field.

Examples

\dontrun{

#Make dataObject
dataObj <- intModel(...)

#Add biasfield for a dataset
dataObj$addBias(datasetNames = 'dataset', biasField = inla.spde2.pcmatern(...))

}

Method updateFormula()

Function to update formulas for a given process.

Usage

dataSDM$updateFormula(
  datasetName = NULL,
  Points = TRUE,
  speciesName = NULL,
  markName = NULL,
  Formula,
  allProcesses = FALSE,
  newFormula,
  ...
)

Arguments

datasetName

Name of the dataset to change the formula for.

Points

Logical: should the formula be changed for the points. Defaults to TRUE. If FALSE, then markNames needs to be non-null.

speciesName

Name of the species to change the formula for.

markName

Name of the mark to change the formula for.

Formula

An updated formula to give to the process.

allProcesses

Logical argument: if TRUE changes the formulas for all processes in a dataset.

newFormula

Completely change the formula for a process -- primarily used to add non-linear components into the formula. Note: all terms need to be correctly specified here.

Examples

\dontrun{

#Make data object
dataObj <- intModel(...)

#View formulas for a dataset
dataObj$updateFormula(datasetName = 'dataset')
#View formula for a species if speciesName is non-null in intModel
dataObj$updateFormula(datasetName = 'dataset', speciesName = 'species')

#Change formula across a dataset
#Main purpose is to remove terms for a specific process
#So start with full model with intModel()
#And thin terms out with .$updateFormula()
dataObj$updateFormula(datasetName = 'dataset', Formula = ~ . -covariate)
dataObj$updateFormula(datasetName = 'dataset', Formula = ~ . -shared_spatial)

#If you would like to add a special term to the model, use the NewFormula argument
dataObj$updateFormula(datasetName = 'dataset', Formula ~ . + (covariate+1e-5)*scaling)
#And then also add that term to the components
dataObj$changeComponents(newComponent = 'scaling')

 }

Method changeComponents()

Function to add custom components to the integrated modeling.

Usage

dataSDM$changeComponents(addComponent, removeComponent, print = TRUE)

Arguments

addComponent

Component to add to the model.

removeComponent

Component (or name of a component) present in the model which should be removed.

print

Logical: should the updated components be printed. Defaults to TRUE.

Examples

\dontrun{

#Make data object first
dataObj <- intModel(...)

#Change a component
dataObj$changeComponent(addComponent = 'dataset_bias(main = coordinates, model = dataset_bias_field, copy = "shared_spatial)')
#Should update the bias field

#Remove a component
datasetObj$changeComponent(removeComponent = 'shared_spatial')
#Will remove shared_spatial from the components + all the formulas it is in

}

Method priorsFixed()

Function to change priors for the fixed (and possibly random) effects of the model.

Usage

dataSDM$priorsFixed(
  Effect,
  Species = NULL,
  datasetName = NULL,
  mean.linear = 0,
  prec.linear = 0.001
)

Arguments

Effect

Name of the fixed effect covariate to change the prior for.

Species

Name of the species for which the prior should change. Defaults to NULL which will change the prior for all species added to the model.

datasetName

Name of the dataset for which the prior of the intercept should change (if fixedEffect = 'intercept'). Defaults to NULL which will change the prior effect of the intercepts for all the datasets in the model.

mean.linear

Mean value for the prior of the fixed effect. Defaults to 0.

prec.linear

Precision value for the prior of the fixed effect. Defaults to 0.001.

Examples

\dontrun{

#Make data object
dataObj <- intModel(...)

#Change priors for the fixed effect
dataObj$priorsFixed(Effect = 'covariate', mean.linear = 2, prec.linear = 0.05)

#Change priors for intercept
dataObj$priorsFixed(Effect = 'intercept', datasetName = 'datasetname',
                    mean.linear = 2, prec.linear = 0.005)

}

Method specifySpatial()

Function to specify the random fields in the model using PC priors for the parameters.

Usage

dataSDM$specifySpatial(
  sharedSpatial = FALSE,
  Species,
  Mark,
  Bias,
  PC = TRUE,
  Remove = FALSE,
  ...
)

Arguments

sharedSpatial

Logical: specify the shared spatial field in the model. Defaults to FALSE.

Species

Name of the species's spatial field to be specified.

Mark

Name of the mark's spatial field to be specified.

Bias

Name of the dataset's bias field to be specified.

PC

Logical: should the Matern model be specified with pc priors. Defaults to TRUE.

Remove

Logical: should the spatial field be removed. Requires one of sharedSpatial, species, mark or bias to be non-missing.

...

Additional arguments used by INLA's inla.spde2.pcmatern or inla.spde2.matern function.

Examples

\dontrun{

#Create data object
dataObj <- intModel(...)

#Specify spatial field
dataObj$specifySpatial(sharedSpatial = TRUE, PC = TRUE, prior.range = c(1,0.001), prior.sigma = c(1,0.001))
#Specify for a species
dataObj$specifySpatial(species = 'species', alpha = 1, PC = FALSE)

#Remove a spatial field
dataObj$specifySpatial(species = 'species', remove = TRUE)
#Which will also update the components and relevant formulas

}

Method changeLink()

Function used to change the link function for a given process.

Usage

dataSDM$changeLink(datasetName, Species, Mark, Link, ...)

Arguments

datasetName

Name of the dataset for which the link function needs to be changed.

Species

Name of the species for which the link function needs to be changed.

Mark

Name of the mark for which the link function needs to be changed.

Link

Name of the link function to add to the process. If missing, will print the link function of the specified dataset.

...

Not used

Examples

\dontrun{

#Create data object
dataObj <- intModel(...)

#Print link function for a process

dataObj$changeLink(Dataset = Dataset, Species = Species)

#Change link function
dataObj$changeLink(Dataset = Dataset, Species = Species,
                   Link = "log")

}

Method spatialBlock()

Function to spatially block the datasets. See the spatialBlock function from blockCV for spatial blocking.

Usage

dataSDM$spatialBlock(k, rows, cols, plot = FALSE, ...)

Arguments

k

Number of cross-validation folds.

rows

Integer value by which the area is divided into latitudinal bins.

cols

Integer value by which the area is divided into longitudinal bins.

plot

Plot the cross-validation folds. Defaults to FALSE.

...

Extra arguments used by blockCV's spatialBlock

Examples

\dontrun{

#Make data object
dataObj <- intModel(...)

#Block the points spatially
dataObj$spatialBlock(k = 5, rows = 2, cols = 5, plot = TRUE)

}

Method new()

Usage

dataSDM$new(
  coordinates,
  projection,
  Inlamesh,
  initialnames,
  responsecounts,
  responsepa,
  marksnames,
  marksfamily,
  pointcovariates,
  trialspa,
  trialsmarks,
  speciesname,
  marksspatial,
  spatial,
  intercepts,
  spatialcovariates,
  marksintercepts,
  boundary,
  ips,
  temporal,
  temporalmodel,
  speciesspatial
)

Method samplingBias()

Usage

dataSDM$samplingBias(...)

Examples

## ------------------------------------------------
## Method `dataSDM$plot`
## ------------------------------------------------

if (FALSE) {

#Make data object
dataObj <- intModel(...)

#Plot a specific datasets
dataObj$plot(Datasets = c('dataset1', 'dataset2'))

#Plot all datasets
dataObj$plot()

#If speciesName specified in intModel, plot species
dataObj(Species = TRUE)

}

## ------------------------------------------------
## Method `dataSDM$addData`
## ------------------------------------------------

if (FALSE) {

#Make data object with specified components
dataObj <- intModel(...)

#Add new data with non-standardized names
dataObj$addData(dataset, responseCounts = responseCounts,
                responsePA = responsePA, trialsPA = trialsPA,
                markNames = markNames, markFamily = markFamily,
                pointCovariates = pointCovariates,
                speciesName = speciesName, Coordinates = Coordinates)

}

## ------------------------------------------------
## Method `dataSDM$addBias`
## ------------------------------------------------

if (FALSE) {

#Make dataObject
dataObj <- intModel(...)

#Add biasfield for a dataset
dataObj$addBias(datasetNames = 'dataset', biasField = inla.spde2.pcmatern(...))

}

## ------------------------------------------------
## Method `dataSDM$updateFormula`
## ------------------------------------------------

if (FALSE) {

#Make data object
dataObj <- intModel(...)

#View formulas for a dataset
dataObj$updateFormula(datasetName = 'dataset')
#View formula for a species if speciesName is non-null in intModel
dataObj$updateFormula(datasetName = 'dataset', speciesName = 'species')

#Change formula across a dataset
#Main purpose is to remove terms for a specific process
#So start with full model with intModel()
#And thin terms out with .$updateFormula()
dataObj$updateFormula(datasetName = 'dataset', Formula = ~ . -covariate)
dataObj$updateFormula(datasetName = 'dataset', Formula = ~ . -shared_spatial)

#If you would like to add a special term to the model, use the NewFormula argument
dataObj$updateFormula(datasetName = 'dataset', Formula ~ . + (covariate+1e-5)*scaling)
#And then also add that term to the components
dataObj$changeComponents(newComponent = 'scaling')

 }

## ------------------------------------------------
## Method `dataSDM$changeComponents`
## ------------------------------------------------

if (FALSE) {

#Make data object first
dataObj <- intModel(...)

#Change a component
dataObj$changeComponent(addComponent = 'dataset_bias(main = coordinates, model = dataset_bias_field, copy = "shared_spatial)')
#Should update the bias field

#Remove a component
datasetObj$changeComponent(removeComponent = 'shared_spatial')
#Will remove shared_spatial from the components + all the formulas it is in

}

## ------------------------------------------------
## Method `dataSDM$priorsFixed`
## ------------------------------------------------

if (FALSE) {

#Make data object
dataObj <- intModel(...)

#Change priors for the fixed effect
dataObj$priorsFixed(Effect = 'covariate', mean.linear = 2, prec.linear = 0.05)

#Change priors for intercept
dataObj$priorsFixed(Effect = 'intercept', datasetName = 'datasetname',
                    mean.linear = 2, prec.linear = 0.005)

}

## ------------------------------------------------
## Method `dataSDM$specifySpatial`
## ------------------------------------------------

if (FALSE) {

#Create data object
dataObj <- intModel(...)

#Specify spatial field
dataObj$specifySpatial(sharedSpatial = TRUE, PC = TRUE, prior.range = c(1,0.001), prior.sigma = c(1,0.001))
#Specify for a species
dataObj$specifySpatial(species = 'species', alpha = 1, PC = FALSE)

#Remove a spatial field
dataObj$specifySpatial(species = 'species', remove = TRUE)
#Which will also update the components and relevant formulas

}

## ------------------------------------------------
## Method `dataSDM$changeLink`
## ------------------------------------------------

if (FALSE) {

#Create data object
dataObj <- intModel(...)

#Print link function for a process

dataObj$changeLink(Dataset = Dataset, Species = Species)

#Change link function
dataObj$changeLink(Dataset = Dataset, Species = Species,
                   Link = "log")

}

## ------------------------------------------------
## Method `dataSDM$spatialBlock`
## ------------------------------------------------

if (FALSE) {

#Make data object
dataObj <- intModel(...)

#Block the points spatially
dataObj$spatialBlock(k = 5, rows = 2, cols = 5, plot = TRUE)

}