Predict for bru_sdm

Predict for bru_sdm

# S3 method for bruSDM
predict(
  object,
  data = NULL,
  formula = NULL,
  mesh = NULL,
  mask = NULL,
  temporal = FALSE,
  covariates = NULL,
  spatial = FALSE,
  intercepts = FALSE,
  datasets = NULL,
  species = NULL,
  biasfield = FALSE,
  biasnames = NULL,
  predictor = FALSE,
  fun = "exp",
  ...
)

Arguments

object

A bru_sdm object.

data

Data containing points of the map with which to predict on. May be NULL if one of mesh or mask is NULL.

formula

Formula to predict. May be NULL if other arguments: covariates, spatial, intercepts are not NULL.

mesh

An inla.mesh object

mask

A mask of the study background. Defaults to NULL.

temporal

Make predictions for the temporal component of the model.

covariates

Name of covariates to predict.

spatial

Include spatial effects in prediction.

intercepts

Include intercept in prediction.

datasets

Names of the datasets to include intercept term. Default of NULL results in all datasets being predicted.

species

Names of the species to predict. Default of NULL results in all species being predicted.

biasfield

Include bias field in prediction.

biasnames

Names of the datasets to include bias term. Defaults to NULL.

predictor

Should all terms run in the linear predictor be included. Defaults to FALSE.

fun

Function used to predict. Set to 'linear' if effects on the linear scale are desired.

...

Additional arguments used by the inlabru predict function.

Value

A list of inlabru predict objects.

Examples

if (FALSE) {

#Make data object
dataObj <- intModel(...)

##Run model
intModel <- runModel(dataObj, ...)

#Predict
predictions <- predict(object = intModel, mask = spatialPolygons,
                       mesh = inlaMesh, spatial = TRUE, covariates = 'covariates',
                       fun = 'linear')
                       
#Predict linear predictor
predictions <- predict(object = intModel, data = spatialPolygons,
                      mesh = inlaMesh, predictor = TRUE)
                      
#Predict bias field
predictions <- predict(object = intModel, data = spatialPolygons,
                      mesh = inlaMesh, bias = 'datasetName')
                       

}